requirements.txt
playwright
beautifulsoup4
loguru
python-dotenv
requests
Pillow
twocaptcha
pytest
pytest-mock
pandas
